updated with user edit form fix (#1230)
